Fix for MGGA and k-points.  Need to add a test also ...
git-svn-id: https://svn.fysik.dtu.dk/projects/gpaw/trunk@7535 6419b465-d902-0410-997c-be2266baf723

+++ gpaw/xc/mgga.py
@@ -25,7 +25,8 @@ class MGGA(GGA):
         self.dedtaut_sG = None
         self.restrict = hamiltonian.restrictor.apply
         self.interpolate = density.interpolator.apply
-        self.taugrad_v = [Gradient(wfs.gd, v, allocate=True).apply
+        self.taugrad_v = [Gradient(wfs.gd, v, dtype=wfs.dtype,
+                                   allocate=True).apply
                           for v in range(3)]
 
     def set_positions(self, spos_ac):
@@ -55,7 +56,7 @@ class MGGA(GGA):
                                                
     def apply_orbital_dependent_hamiltonian(self, kpt, psit_xG,
                                             Htpsit_xG, dH_asp):
-        a_G = self.wfs.gd.empty()
+        a_G = self.wfs.gd.empty(dtype=psit_xG.dtype)
         for psit_G, Htpsit_G in zip(psit_xG, Htpsit_xG):
             for v in range(3):
                 self.taugrad_v[v](psit_G, a_G, kpt.phase_cd)
